:recycle: refactor(domain): type 및 interface를 정의한 파일명 변경
